Project/requirements.txt
django 
djangorestframework
requests
Pillow
djangorestframework-simplejwt
openai
g4f[all]
curl_cffi
djangorestframework-authtoken